requirements.txt
dlib==19.19.0
Pillow==7.0.0
scikit-image==0.16.2
scipy==1.4.1
